// src/api/weather-forecast-service/models/WeatherForecastResponse.ts
export interface WeatherForecastResponse {
    /**
     * 
     * @type {number}
     * @memberof WeatherForecastResponse
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof WeatherForecastResponse
     */
    location?: string;
    /**
     * 
     * @type {number}
     * @memberof WeatherForecastResponse
     */
    latitude?: number;
    /**
     * 
     * @type {number}
     * @memberof WeatherForecastResponse
     */
    longitude?: number;
    /**
     * 
     * @type {string}
     * @memberof WeatherForecastResponse
     */
    timezone?: string;
    /**
     * 
     * @type {Date}
     * @memberof WeatherForecastResponse
     */
    lastUpdated?: Date;
    /**
     * 
     * @type {Array<DailyForecastResponse>}
     * @memberof WeatherForecastResponse
     */
    dailyForecasts?: Array<DailyForecastResponse>;
    /**
     * 
     * @type {Array<HourlyForecastResponse>}
     * @memberof WeatherForecastResponse
     */
    hourlyForecasts?: Array<HourlyForecastResponse>;
}